[FIX] setup: Redirect to installer if no prefs table found (would be nicer to display the db error but nothing is initialised at this point.) Happens if you edit the local.php manually and point it at a nonexistent database, for instance)
git-svn-id: https://svn.code.sf.net/p/tikiwiki/code/trunk@35445 b456876b-0849-0410-b77d-98878d47e9d5

+++ tiki-setup_base.php
@@ -62,6 +62,12 @@ $needed_prefs = array(
 	'min_pass_length' => 5,
 	'pass_chr_special' => 'n'
 );
+// check that tiki_preferences is there
+if (empty($tikilib->query("SHOW TABLES LIKE 'tiki_preferences'")->numrows)) {
+	// smarty not initialised at this point to do a polite message, sadly
+	header('location: tiki-install.php');
+	exit;
+}
 $tikilib->get_preferences($needed_prefs, true, true);
 if (!isset($prefs['lastUpdatePrefs']) || $prefs['lastUpdatePrefs'] == - 1) {
 	$tikilib->query('delete from `tiki_preferences` where `name`=?', array('lastUpdatePrefs'));
